test: reorganize subgraph E2E tests into domain-organized directory
Restructure 17 flat spec files into 10 domain-organized files under
browser_tests/tests/subgraph/, grouped by behavioral concern:

- subgraphSlots: I/O CRUD, rename, alignment, slot position
- subgraphCrud: create, delete, copy, unpack
- subgraphOperations: copy/paste and undo/redo inside subgraphs
- subgraphNavigation: breadcrumb, viewport, hotkeys, progress state
- subgraphPromotion: auto-promote, visibility, reactivity, context menu
- subgraphPromotionDom: DOM widget persistence, cleanup, positioning
- subgraphSerialization: hydration, round-trip, legacy formats, ID remap
- subgraphNested: configure order, duplicate names, pack values
- subgraphLifecycle: source removal cleanup, pseudo-preview lifecycle
- subgraphSearch: search aliases, description, persistence

No test logic changes. All 96 test cases preserved with original names.
Consistent camelCase naming. Follows existing subdirectory pattern
(vueNodes/, sidebar/).

Amp-Thread-ID: https://ampcode.com/threads/T-019d3709-4138-737c-bad2-7220f6fd6fce
